+++ Tests/Unit/MSFT_ADServicePrincipalName.Tests.ps1
@@ -105,13 +105,13 @@ try
                 It 'Should return false for present' {
 
                     $result = Test-TargetResource -Ensure 'Present' @testDefaultParameters
-                    $result | Should -Be $false
+                    $result | Should -BeFalse
                 }
 
                 It 'Should return true for absent' {
 
                     $result = Test-TargetResource -Ensure 'Absent' @testDefaultParameters
-                    $result | Should -Be $true
+                    $result | Should -BeTrue
                 }
             }
 
@@ -124,13 +124,13 @@ try
                 It 'Should return true for present' {
 
                     $result = Test-TargetResource -Ensure 'Present' @testDefaultParameters
-                    $result | Should -Be $true
+                    $result | Should -BeTrue
                 }
 
                 It 'Should return false for absent' {
 
                     $result = Test-TargetResource -Ensure 'Absent' @testDefaultParameters
-                    $result | Should -Be $false
+                    $result | Should -BeFalse
                 }
             }
 
@@ -143,13 +143,13 @@ try
                 It 'Should return false for present' {
 
                     $result = Test-TargetResource -Ensure 'Present' @testDefaultParameters
-                    $result | Should -Be $false
+                    $result | Should -BeFalse
                 }
 
                 It 'Should return false for absent' {
 
                     $result = Test-TargetResource -Ensure 'Absent' @testDefaultParameters
-                    $result | Should -Be $false
+                    $result | Should -BeFalse
                 }
             }
 
@@ -163,13 +163,13 @@ try
                 It 'Should return false for present' {
 
                     $result = Test-TargetResource -Ensure 'Present' @testDefaultParameters
-                    $result | Should -Be $false
+                    $result | Should -BeFalse
                 }
 
                 It 'Should return false for absent' {
 
                     $result = Test-TargetResource -Ensure 'Absent' @testDefaultParameters
-                    $result | Should -Be $false
+                    $result | Should -BeFalse
                 }
             }